Decompose tests
