+++ README.md
@@ -5,13 +5,11 @@
 
 Load BIOM tables and sample metadata from a lot of studies into Redis to facilitate search using observation IDs or metadata.
 
-It is safe to say the interfaces are not stable.
-
 # Installation
 
-Needs Redis (tested with 3.2.6) and Webdis (just clone the repo). Also depends on BIOM (tested on 2.1.5-dev), Pandas (tested on 0.19.0), Click (required >= 6.7) and scipy (whatever BIOM is happy with).
+Needs Redis (tested with 3.2.6) and Webdis (just clone the repo). Also depends on BIOM (tested on 2.1.5), Pandas (tested on 0.19.0), Click (required >= 6.7) and scipy (whatever BIOM is happy with).
 
-# Load some stuff
+# Load some data
 
 To make use of this cache, we need to load things. Loading can be done in parallel. First, we'll load up metadata. This will create keys in Redis which describe all of the columns associated with a sample (e.g., `metadata:categories:<sample_id>`, hash buckets for each category and sample combination (e.g., `metadata:category:<category_name>` as the hash and `<sample_id>` as the field), a set of all known categories (e.g., `metadata:categories-represented`), and a set of all known sample IDs (e.g., `metadata:samples-represented`):
 
@@ -25,11 +23,11 @@ Next, we'll load up associations between every single observation in a BIOM tabl
 
 	$ redbiom admin load-observations --context deblur-100nt --table /path/to/biom/table.biom
 
-For icing on the cake, lets also load up all of the BIOM table data because why not. We'll only store the non-zero values, and we'll encode the sample data into something simple so that it goes in as just a string to Redis. Important: we only support storing count data right now, not floating point. The keys created are of the form `<context_name>:data:<sample_id>`. To reduce space, we reindex the observation IDs as things like sOTUs tend to be very long in name. The mapping is stable over all tables loaded (ie the same observation has the same index), and the index is stored as a JSON object under the key `<context_name>:__observation_index`. Because we need to update the index, this operation cannot be done in parallel however the code is setup with a redis-based mutex so it's okay to queue up multiple loads.
+Last, let's load up all of the BIOM table data. We'll only store the non-zero values, and we'll encode the sample data into something simple so that it goes in as just a string to Redis. Important: we only support storing count data right now, not floating point. The keys created are of the form `<context_name>:data:<sample_id>`. To reduce space, we reindex the observation IDs as things like sOTUs tend to be very long in name. The mapping is stable over all tables loaded (ie the same observation has the same index), and the index is stored as a JSON object under the key `<context_name>:__observation_index`. Because we need to update the index, this operation cannot be done in parallel however the code is setup with a redis-based mutex so it's okay to queue up multiple loads.
 
 	$ redbiom load-sample-data --context deblur-100nt --table /path/to/biom/table.biom
 
-# Get you some stuff
+# Query for content
 
 Now that things are loaded, we can search for stuff. Let's say you have a few OTUs of interest, and you are curious about what other samples they exist in. You can find that out with:
